+++ app/controllers/p_controller.rb
@@ -1,9 +1,10 @@
 class PController < ApplicationController
-  def index
+  def browse
     @podcasts = Podcast.all
   end
 
   def show
     @podcast = Podcast.find(params[:id])
+    @episodes = @podcast.episodes
   end
 end
