+++ app/views/p/browse.html.erb
@@ -1,0 +1,30 @@
+<header class="flex justify-between bg-gray-700">
+  <div class="flex m-6 justify-center md:justify-start">
+    <%= image_tag("logo", size: "180x180") %>
+  </div>
+  <form class="search-form border hover:border-blue-600 rounded flex my-auto mr-10">
+    <button class="search-btn focus:outline-none w-10 bg-white rounded-l text-gray-500 border-none p-2"><i class="fas fa-search"></i></button><input type="text" class="search-input focus:outline-none w-72 bg-white rounded-r border-none p-2" placeholder="搜尋 Podcast">
+  </form>
+</header>
+
+<section class="bg-gray-700 h-screen">
+  <div class="flex items-center justify-center">
+    <div class="grid grid-cols-1 gap-6 sm:grid-cols-2 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
+      <% @podcasts.each do |podcast| %>
+        <div class="relative bg-white py-6 px-6 rounded-3xl w-80 my-4 shadow-xl transform hover:-translate-y-1 hover:scale-100">
+          <div class="flex">
+            <div class="p-3"><img src="https://picsum.photos/30/30/?random=1"></div>
+            <div>
+              <h3 class="font-black duration-500 hover:text-blue-500"><%= link_to podcast.name, p_path(podcast.id) %></h3>
+              <p class="font-light"><%= podcast.artist %></p>
+            </div>
+          </div>
+          <div class="border-t-2"></div>
+          <div class="flex justify-center">
+          </div>
+        </div>
+      <% end %>
+    </div>
+  </div>
+</section>
+
