--- app/views/p/index.html.erb
@@ -1,3 +1,0 @@
-<h1>P#index</h1>
-<p>Find me in app/views/p/index.html.erb</p>
-
